src/test/java/BibliotecaTest.java
import com.teste.implementabiblioteca.Biblioteca;
import com.teste.implementabiblioteca.Endereco;
import com.teste.implementabiblioteca.Livro;
import org.junit.jupiter.api.Test;
import java.util.ArrayList;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

class BibliotecaTest {

    @Test
    public void Test_Construtor() {

        List<Livro> livro_atual = new ArrayList<>();

        Endereco novo_endereco = new Endereco
                ("Afonso", 25, "Centro", "Maringá", "Paraná", 1);

        Biblioteca nova_biblioteca = new Biblioteca("Sarandi", livro_atual, novo_endereco, 1);
        assertEquals("Sarandi",nova_biblioteca.getNome_da_biblioteca());
        assertEquals(livro_atual,nova_biblioteca.getLista_de_livro());
        assertEquals(novo_endereco,nova_biblioteca.getEndereco());
        assertEquals(1,nova_biblioteca.getId_biblioteca());

    }

}